+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 12.4.0 - 2022-05-10
 
 ### Added
-- Added ability to use `--extraBrowserCapabilities` in bedrock-auto command.
+- Additional webdriver browser capabilities can be passed via the `--extraBrowserCapabilities` argument in the bedrock-auto command.
+- Verbose logging mode can be passed via the `--verbose` argument in bedrock and bedrock-auto
 
 ## 12.3.2 - 2022-01-04